Remove initialize method from FieldSerializer module
This causes problems because this initialize method can clash with the
initialize method in other classes, causing unexpected errors depending
on the order that things are defined.

It seems like the best thing to do is remove this entirely. Then it's up
to other libraries (such as `scraped_page`) to provide a proper
initializer.